fix snackbar position and and remove redundent code in details list bottom sheet

+++ feature/home/presentation/src/main/java/com/sanaa/presentation/bottomsheet/saveToListBottomsheet/SaveToListViewModel.kt
@@ -74,8 +74,8 @@ class SaveToListViewModel @Inject constructor(
         mediaId: Int
     ){
         manageSavedListItemsUseCase.addMovieToSavedList(
-            listId = selectedListId.toInt(),
-            movieId = mediaId.toInt()
+            listId = selectedListId,
+            movieId = mediaId
         )
     }
