+++ feature/media_details/presentation/src/main/java/com/sanaa/presentation/bottomsheets/saveToListBottomsheet/SaveToListsInteractionListener.kt
@@ -1,7 +1,7 @@
 package com.sanaa.presentation.bottomsheets.saveToListBottomsheet
 
 interface SaveToListBottomSheetInteractionListener {
-    fun onPlaylistClick(listId: Long)
+    fun onPlaylistClick(listId: Int)
     fun onAddClick()
     fun onSnackBarDismiss()
     fun onCreateNewListClick()
